optimized all imports and formatted code as per new schema

+++ src/Database/Migrations/2018_04_04_125846_create_dz_model_details_table.php
@@ -6,23 +6,25 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateDzModelDetailsTable extends Migration {
+class CreateDzModelDetailsTable extends Migration
+{
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up () {
+    public function up ()
+    {
         Schema::create('dz_model_details', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
             $routeTable = ( new Route() )->getTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
             $table->string('name');
             $table->string('description')->nullable();
 
-            $table->unsignedBigInteger('route_id')->nullable();
+            $table->unsignedInteger('route_id')->nullable();
 
             $table->string('namespace')->nullable();
             $table->string('table_name')->nullable();
@@ -32,8 +34,8 @@ class CreateDzModelDetailsTable extends Migration {
             $table->string('model_hash');
             $table->string('route_name')->nullable();
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('route_id')->references('id')->on($routeTable);
             $table->foreign('created_by')->references('id')->on($userTable);
@@ -51,7 +53,8 @@ class CreateDzModelDetailsTable extends Migration {
      *
      * @return void
      */
-    public function down () {
+    public function down ()
+    {
         Schema::dropIfExists('dz_model_details');
     }
 }
