+++ src/Database/Migrations/2019_03_07_052336_create_dz_observer_events_table.php
@@ -8,27 +8,29 @@ use Illuminate\Support\Facades\Schema;
 /**
  * Class CreateDzObserverEventsTable
  */
-class CreateDzObserverEventsTable extends Migration {
+class CreateDzObserverEventsTable extends Migration
+{
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up () {
+    public function up ()
+    {
         Schema::create('dz_observer_events', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
             $table->string('model_hash')->nullable();
-            $table->unsignedBigInteger('record_id');
+            $table->unsignedInteger('record_id');
 
             $table->longText('data')->nullable();
 
             $table->dateTime('processed_at')->nullable();
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
@@ -43,7 +45,8 @@ class CreateDzObserverEventsTable extends Migration {
      *
      * @return void
      */
-    public function down () {
+    public function down ()
+    {
         Schema::dropIfExists('dz_observer_events');
     }
 }
