+++ src/Database/Migrations/2019_03_20_090404_add_dz_device_tokens_table.php
@@ -5,28 +5,30 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AddDzDeviceTokensTable extends Migration {
+class AddDzDeviceTokensTable extends Migration
+{
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up () {
+    public function up ()
+    {
         Schema::create('dz_device_tokens', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('user_id')->nullable();
+            $table->increments('id');
+            $table->unsignedInteger('user_id')->nullable();
 
             $table->string('token');
             $table->dateTime('last_access_time')->nullable();
             $table->string('version')->nullable();
 
-            $table->unsignedBigInteger('token_source_id')->nullable();
-            $table->unsignedBigInteger('platform_id')->nullable();
+            $table->unsignedInteger('token_source_id')->nullable();
+            $table->unsignedInteger('platform_id')->nullable();
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('user_id')->references('id')->on($userTable);
 
@@ -46,7 +48,8 @@ class AddDzDeviceTokensTable extends Migration {
      *
      * @return void
      */
-    public function down () {
+    public function down ()
+    {
         Schema::dropIfExists('dz_device_tokens');
     }
 }
