+++ src/Database/Migrations/2019_04_29_061900_add_dz_server_deployments_table.php
@@ -5,17 +5,19 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AddDzServerDeploymentsTable extends Migration {
+class AddDzServerDeploymentsTable extends Migration
+{
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up () {
+    public function up ()
+    {
         Schema::create('dz_server_deployments', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
             $table->string('name')->nullable();
 
@@ -34,8 +36,8 @@ class AddDzServerDeploymentsTable extends Migration {
             $table->boolean('active')->default(true);
             $table->dateTime('last_ping_time')->nullable();
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
@@ -50,7 +52,8 @@ class AddDzServerDeploymentsTable extends Migration {
      *
      * @return void
      */
-    public function down () {
+    public function down ()
+    {
         Schema::dropIfExists('dz_server_deployments');
     }
 }
